+++ src/Controller/EditorController.php
@@ -3,8 +3,12 @@
 namespace App\Controller;
 
 use App\Entity\Editor;
+use App\Entity\EditorSearch;
+use App\Form\EditorSearchType;
 use App\Form\EditorType;
+use App\Repository\BookRepository;
 use App\Repository\EditorRepository;
+use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -18,10 +22,21 @@ class EditorController extends AbstractController
     /**
      * @Route("/List_of_Editor", name="editor_index", methods={"GET"})
      */
-    public function index(EditorRepository $editorRepository): Response
+    public function index(EditorRepository $editorRepository,
+                          PaginatorInterface $paginator,
+                          Request $request): Response
     {
+        $search = new EditorSearch();
+        $form = $this->createForm(EditorSearchType::class, $search);
+        $form->handleRequest($request);
+        $editors = $paginator->paginate(
+            $editorRepository->findEditor($search),
+            $request->query->getInt('page', 1),
+            10
+        );
         return $this->render('editor/index.html.twig', [
-            'editors' => $editorRepository->findAll(),
+            'editors' => $editors,
+            'form' => $form->createView()
         ]);
     }
 
@@ -39,7 +54,7 @@ class EditorController extends AbstractController
             $entityManager->persist($editor);
             $entityManager->flush();
 
-            return $this->redirectToRoute('editor_index');
+            return $this->redirectToRoute('editor_show', ['id' => $editor->getId()]);
         }
 
         return $this->render('editor/new.html.twig', [
@@ -59,7 +74,7 @@ class EditorController extends AbstractController
     }
 
     /**
-     * @Route("/editAuthor/{id}", name="editor_edit", methods={"GET","POST"})
+     * @Route("/editEditor/{id}", name="editor_edit", methods={"GET","POST"})
      */
     public function edit(Request $request, Editor $editor): Response
     {
@@ -69,7 +84,9 @@ class EditorController extends AbstractController
         if ($form->isSubmitted() && $form->isValid()) {
             $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('editor_index');
+            return $this->redirectToRoute('editor_show', [
+                'id' => $editor->getId()
+            ]);
         }
 
         return $this->render('editor/edit.html.twig', [
@@ -81,14 +98,21 @@ class EditorController extends AbstractController
     /**
      * @Route("/delete_editor/{id}", name="editor_delete", methods={"DELETE"})
      */
-    public function delete(Request $request, Editor $editor): Response
+    public function delete(Request $request, Editor $editor, BookRepository $bookRepository): Response
     {
         if ($this->isCsrfTokenValid('delete'.$editor->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
+            $book = $bookRepository->findBy(array(
+                'editor' => $editor
+            ));
+            if($book != null){
+                $this->addFlash('failed', 'Cet éditeur a publier certains livres, impossible de supprimer !');
+                return $this->redirectToRoute('editor_show', ['id' => $editor->getId()]);
+            }
             $entityManager->remove($editor);
             $entityManager->flush();
         }
-
+        $this->addFlash('success', "Supprimer avec succès !");
         return $this->redirectToRoute('editor_index');
     }
 }
